Fix cairo_font_face_t leak in CairoFontEngine
Remove the external font cache from CairoFontEngine.  Fixing the leak
would require keeping track of which document is using each external
font so fonts can be removed from the cache when the document is
closed. As fonts are already cached in fontCache, and the external
font caching did not exist prior to the the CairoFontEngine refactor,
the marginal benefit of sharing external fonts between documents is
not worth the effort.