Annots: partial revert of commit 0b474ffb to except form buttons
The original commit unconditionally creates a new
appearance and thus discards the previous selection
state of a checkbox. As a result, only one checkbox
could be selected at a time.

Given the motivation for commit 0b474ffb was to affect
free text annotations, we here restore the original
behavior for any formButton.

Fixes:
Issue #1504
https://gitlab.gnome.org/GNOME/evince/-/issues/2043