Remove incorrect \n and \r characters at the end of error() strings
Some error() message strings end with \n, and \r\n for one of them.
This is incorrect according to the error() code (in poppler/Error.cc),
where control characters are changed to a printable form; a \n is
added automatically by this function.

This affects at least pdfinfo and pdftotext (and the xpdf PDF viewer),
which output error messages like

Syntax Error: Can't get Fields array<0a>

with a spurious <0a> at the end.

This patch removes these \n and \r; it has been obtained with

  perl -pi -e 's/(error\(.*?)(\\[nr])+/\1/' $(git grep -l 'error(.*\\n')

Signed-off-by: Vincent Lefevre <vincent@vinc17.net>