cairo: prevent infinite loop in functionShadedFill()
Prevent y_step or x_step to be zero, as that would cause
infinite loop. Reproducible with the pdf file posted in
issue #1520

Fixed by setting a minimum step to force upon x_step and
y_step, to avoid approximate or reach infinite loop when
x_step or y_step approximates to or equals zero.

Thanks Wesley Moore for reporting and debugging.

Closes #1520