Remove Qt 5 Android CI
This is based on images by KDE that are about to be removed, as part of
retiring Qt 5 Android CI infrastructure there.

Practically this isn't much of a problem, as Qt 5 is meanwhile lagging
behind several Android SDK versions and thus isn't able to produce APKs
that are compliant with the Google Play requirements anymore.