cairo: Fix JPX transparency regression
Commit 2dcfefb43c from https://bugs.freedesktop.org/93468
fixed images 2) and 3) of file[1] (which are jpeg images
with transparency) from showing garbage to show fine
but with a solid background instead of transparency.

But this fix interferes with our previous commit where
we are adding support to Cairo backend to correctly show
transparency in jpeg files that have it.

So we need to partially revert commit 2dcfefb43c but only
for the Cairo backend, because the Splash backend still
uses this fix to improve some jpeg transparent images by
at least showing them with a solid background.

So we added a new method JPXStream::setSupportJPXtransparency()
so a JPXStream can be informed that the OutputDevice is
capable of handling jpx transparency and so to not perform
the downgrade fix to solid background.

The OutputDevice itself can be marked as handling jpx
transparency by new method OutputDev::supportJPXtransparency()
which the Cairo backend implements as TRUE.

So now file[1] does not regress in Cairo backend, although
display changes from showing with solid background to show
with proper transparency as CairoOutputDev now supports that.

For the Splash backend no display regressions should happen
at all, because any display changes are now guard by
OutpuDev::supportJPXtransparency() which the Splash backend
returns FALSE at this moment.

[1] https://bugs.freedesktop.org/attachment.cgi?id=120624

Part of #1486