respect non-ascii encodings in Launch target file names
PDF2.0 (and also 1.7) - section 7.11.3 specifies the UF field
as a "text string" in the pdf sense, which might be encoded
in one of three encodings UTF16. The encoding was ignored here
before resulting in broken links when (1) the UF field was
actually used and (2) the contents was a marked UTF16 string
that would be interpreted as UTF8.

I considered to adjust the parameters too, but these are not officially
defined as text -either "undefined" or partly as "byte string"- and have
been deprecated in PDF2.0.