Simplify getting the Font dict
We want to both get the Dict object and if it is a Ref, get its ref
because we use it later.

Previously we did lookup the Font Object from the dict without fetching
(NF), and if it was a Ref, we had to fetch to get the actual Dict, and if it
was a Dict then all was good.

New code uses the "new" lookup function that returns the Ref as an
output parameter, so we are doing exactly the same but the code is much
simpler