Put a pgp signature in a pdf file
Allow putting a pgp signature in a pdf file.
This is not fully documented inside the specification, but basically it
is replacing the CMS bundle as described with a pgp bundle.

These signatures are tagged in the g10c namespace
(https://github.com/adobe/pdf-names-list/issues/45) and is only
implemented in the GPGME backend as a 'whatever gpg understands'.

A special difference is that the signature is not padded with zeroes,
but with a gpg comment packet with the exact size to fill the void.