[TextOutputDev] Use std::sort instead of qsort, minor cleanup/fixes
Replacing qsort with std::sort is fairly trivial, but gives some extra
type safety and possibly some speedup due to inlining.

The stricter type checks require to change the argument types of e.g.
TextLineFrag::cmp* methods, but also gets rid of the type cast
(to non-const! TextLineFrag*), likewise for the other classes.

The return type change may see deliberate, but is necessary as both
+1 and -1 are evaluated as true, and std::sort requires a compare
function which returns bool. Note, the order is indeterminate in case
both elements are compared as same, but this happens with both
std::sort and qsort (man qsort: "If two members compare as equal,
their order in the sorted array is undefined.").

There are some included changes which are less obvious:

- std::sort(nullptr, nullptr + 0, ...) is safe (no dereference), remove
  the extra nullptr check (`if (blocks)`).
- No need for the immediate `wordArray` in TextWordList, just sort the
  std::vector<TextWord *> words.
- TextBlock::cmpYXPrimaryRot is unused, remove it.