[Splash] Reduce required work for path pruning
For long path, quite some amount of work is spent if the bounding box
of all points intersects with the clipping rectangle, and if there
is some intersection the path is processed further.

For this check, it is often sufficient to only check one, or two points.
Only if both points are outside (which is often false) the remaining
points should be checked.

CCBUG: #1555