[SplashXPathScanner] Optimize count check
The count is incremented/decremented each time a segment crosses a
scanline boundary. If a segment starts/ends on a boundary, the count
has to be changed exactly once, i.e. either on the upper or lower edge of
the segment, but which one is arbitray.

Changing this from the lower to the upper edge simplifies the condition,
as (y <= segYMax) is always true.