Fix pdfcairo's document unit for SVG output
Using cairo 1.17.6 and later,
pdftocairo outputted SVGs with the wrong width and height.

Starting with cairo 1.17.6,
the default SVG units have been changed from pt to user units.
https://gitlab.freedesktop.org/cairo/cairo/-/issues/545

pdftocairo did not set the SVG unit,
and the change in cairo's default
caused pdfcairo's SVG output to be incorrect.

This commit makes pdftocairo set the SVG unit to pt.
So, pdftocairo outputs SVGs that have the correct width and height.