HtmlString: use getCurText[XY]() instead of getCur[XY]()
The fix for #1602 that added curTextX() and curTextY() forgot to adjust
HtmlOutputDev to use the new variables as well, leading to text
coordinates being broken in pdftohtml output. Change HtmlString to use
curTextX() and curTextY() so that it works with the new way of handling
text coordinates.

Fixes #1617
Fixes #1622
Fixes 71bf5552d448a6fdb666f2b61764b61ca197617d