TextOutputDev: Use dump() to implement getText()
As all missing features were added to the getText(...) method, it
is possible to leverage it for the getText(...) implementation,
and thus to remove a large amount of mostly duplicate code.

This also adds support for retrieving the text in "flow" order, while
formerly only PhysicalLayout and raw (document stream) order were
possible.

This also fixes differences between pdftotext and the Qt bindings for
some PhysicalLayout cases, i.e. insertion of blank lines for empty page
regions or spaces between paragraphs.

Fixes poppler#1476.