FormWidgetSignature: ensure V key is created in Field dict
As per PDF spec, the value key "V" for a Signature field
must be created in the Field dict, not in the Widget
Annotation dict.

The bug happened when the Field had the Widget annot as
a child, in contrast to having the Widget keys merged
in the Field dict which is also possible.

Fixes #1640 along with the next commit