FormField: fix Widget child mistakenly added as non-terminal Field
The FormField constructor is the place where we lookup the 'Kids'
array to find possible non-terminal Fields, which are added to
the 'children' vector, and also possible Widget annotation's which
are added to the 'widgets' vector.

Issue #1640 exposes a case where a Widget annot (which was the solely
child of the FormField) was mistakenly added to the 'children' vector
instead of the 'widgets' one. This commit fixes that.

The buggy logic (prior to this commit) was based on PDF spec < 1.7
where Widget Annotations were not allowed to have a "Parent" entry,
but PDF 1.7 spec introduced the "Parent" entry for Widget Annotations
when it is the child of a Field (see Table 188 in 12.5.6.19 section).

Fixes #1640 along with prior commit