Annot.cc: partially revert e674ca64 to fix #1642 regression
The refactor in commit e674ca64 deleted the code in
AnnotWidget::draw() that checked if ZaDb font was
automatically added by AnnotAppearanceBuilder::drawText()
(after it could not find it in the document Resources)
and if that was the case then AnnotWidget::draw() added
a fake ZaDb font to the Gfx Resources prior to calling
gfx->drawAnnot().

The deletion of that code was causing pdf form files
exported from LibreOffice which had checkbuttons and
radiobuttons to not display them as checked after
clicking them.

So this commit brings back that code so those form
buttons display fine again.

Fixes #1642